refactor(atom/input): file restructuring + compliant
