+++ components/atom/input/src/Label/index.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import './styles.scss'
-
 const InputLabel = ({name, label, children}) => (
   <label htmlFor={name}>
     {children}
