+++ components/atom/input/src/Label/index.scss
@@ -7,4 +7,3 @@ label {
   margin-bottom: $m-s;
   margin-right: $m-m;
 }
-
