+++ components/atom/input/src/inputs/TagInput/index.scss
