--- components/atom/input/src/inputs/index.js
@@ -1,9 +1,0 @@
-import MaskInput from './MaskInput'
-import TagInput from './TagInput'
-import PasswordInput from './PasswordInput'
-
-export {
-  MaskInput,
-  TagInput,
-  PasswordInput
-}
